Move shared chart config to shared folder
