+++ apps/web/src/features/projects/views/projects-impacts-comparison/charts/full-time-jobs/ReconversionFullTimeJobsComparisonChart.tsx
@@ -1,6 +1,6 @@
 import * as Highcharts from "highcharts";
 import HighchartsReact from "highcharts-react-official";
-import { sharedChartConfig } from "../sharedChartConfig";
+import { sharedChartConfig } from "../../../shared/sharedChartConfig";
 
 function ReconversionFullTimeJobsComparisonChart() {
   const barChartOptions: Highcharts.Options = {
