+++ apps/web/src/features/projects/views/projects-impacts-comparison/charts/socioeconomic-impacts/SocioEconomicImpactComparison.tsx
@@ -1,6 +1,6 @@
 import * as Highcharts from "highcharts";
 import HighchartsReact from "highcharts-react-official";
-import { sharedChartConfig } from "../sharedChartConfig";
+import { sharedChartConfig } from "../../../shared/sharedChartConfig";
 
 function SocioEconomicImpactComparisonChart() {
   const barChartOptions: Highcharts.Options = {
